pyproject.toml
[project]
name = "eval-by-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "logging>=0.4.9.6",
    "ollama>=0.6.0",
    "pydantic>=2.11.10",
    "requests>=2.32.5",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
    "redis>=6.4.0",
    "schedule>=1.2.2",
]
